ml-service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
PyMuPDF>=1.25.0
pdfplumber==0.11.0
scikit-learn>=1.5.0
nltk==3.8.1
gunicorn==22.0.0
rapidfuzz>=3.6.0
spacy>=3.7.0
